pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mdv"
version = "0.1.0"
description = "Terminal markdown viewer with inline image support for iTerm2"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0",
    "Pillow>=10.0",
]

[project.scripts]
mdv = "mdv.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mdv"]
